requirements.txt
aiohttp==3.8.3
firebase_admin==6.2.0
Flask==2.2.2
nest_asyncio==1.5.7
numpy==1.25.2
pandas==1.5.3
protobuf==4.24.0
python-dotenv==1.0.0
Requests==2.31.0
scikit_learn==1.3.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
gunicorn==20.1.0
connexion[swagger-ui]==2.14.1
